+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Transcoder Changelog
 
+## 1.0.1 - 2017.03.06
+### Added
+- Added `height` and `width` options for resizing the videos
+- Added an `aspectRatio` option to control how aspect ratio scaling is done
+- Added an `letterboxColor` option
+- The `ffmpeg` progress for a particular video is now written out to a `.progress` file
+- Added the `getFileInfo` variable to extract information from a video/audio file
+
+### Fixed
+- Fixed some issues with the lockfile naming
+
 ## 1.0.0 - 2017.03.05
 ### Added
 - Initial release